SAGE-65 prompt engineering
Modified system prompt.

Now, the model generates its own message when it doesn't have context, instead of the system displaying the same message everytime.

This also allows the model to humanly respond to simple queries, such as "Hello", instead of alerting the user that it doesnt have the necessary context to respond to "Hello".